Fix flake8 and mypy findings to unblock CI
These were pre-existing but hidden: CI died on black/isort before
reaching the flake8/mypy steps, so nothing ever ran them until this
session pinned versions and got black+isort green.

flake8: drop an unused import in main.py, wrap a few long lines, and
ignore E203/two vendored-template files' E501 in CI config (black
itself produces the E203-flagged spacing, and the flagged lines are
an unsplittable macOS plist DOCTYPE and raw HTML test fixtures).

mypy (42 errors, all pre-existing):
- classify_fit(score, min_score, good_score) took int thresholds but
  every caller passes float (user-configurable via the dashboard) —
  widened the signature once instead of patching 16 call sites.
- 4 Selenium clients had unannotated `self._driver = None`, which
  mypy pins to NoneType; added Optional[webdriver.Chrome] each place.
- Scheduler.source_map was typed dict[str, Callable[..., None]] but
  handlers return Any and dict is invariant; switched to Mapping
  (covariant) with an Any return, since the return value is never
  used.
- winreg has no attributes in typeshed unless mypy checks against
  win32 — added type: ignore[attr-defined] on the Windows-only
  registry calls in autostart.py, which only run on sys.platform ==
  "win32" anyway.
- 3 telegram endpoints in api.py called TelegramConversations.get()
  after a mutation and returned it directly, but get() can genuinely
  return None; added the same 404/500 guard the function already
  uses elsewhere instead of casting the type away.
- LinkedIn search coerces a bs4 attribute (str | list | None) to str
  before building Job.role, since aria-label is never actually a
  list in practice.
- annotated a local "dry_run"/"applied" status var so AppliedLog.record
  sees the Literal it declares instead of a widened str.

+++ src/job_sources/job_fit.py
@@ -38,7 +38,7 @@ class FitAssessment(BaseModel):
     gaps: list[str] = Field(default_factory=list)
 
 
-def classify_fit(score: int, min_score: int, good_score: int) -> FitTier:
+def classify_fit(score: int, min_score: float, good_score: float) -> FitTier:
     """Три уровня вместо одного порога: ниже min_score — вообще не
     откликаемся; между min_score и good_score — откликаемся, но это
     помечается как слабый матч в статистике; выше good_score — как
